requirements.txt
numpy
opencv-python
pillow
tqdm
pyyaml
matplotlib
pycocotools
torch
torchvision
open-clip-torch
transformers
accelerate
datasets
